requirements.txt
matplotlib==3.1.1
networkx==2.3
nltk==3.4.5
numpy==1.16.5
pandas==0.25.1
pymongo==3.10.1
python-igraph==0.8.0
tweepy==3.8.0